feat(Audit): move most of the code to a trait and implement the contract
